requirements.txt
# HTTP and web scraping
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
certifi>=2023.7.22

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Vector database and embeddings
qdrant-client>=1.7.0
openai>=1.3.0

# Environment and configuration
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0  # Optional, for advanced data analysis

# Development tools (optional)
black>=23.0.0  # Code formatting
pytest>=7.0.0  # Testing
pytest-asyncio>=0.21.0  # Async testing